refactor : jsch 반복되는 코드 수정
