+++ src/main/java/com/ezfarm/ezfarmback/remote/service/RemoteService.java
@@ -4,11 +4,11 @@ import com.ezfarm.ezfarmback.common.exception.CustomException;
 import com.ezfarm.ezfarmback.common.exception.dto.ErrorCode;
 import com.ezfarm.ezfarmback.farm.domain.Farm;
 import com.ezfarm.ezfarmback.farm.domain.FarmRepository;
-import com.ezfarm.ezfarmback.remote.domain.JschObject;
+import com.ezfarm.ezfarmback.common.iot.JschService;
 import com.ezfarm.ezfarmback.remote.domain.OnOff;
 import com.ezfarm.ezfarmback.remote.domain.Remote;
 import com.ezfarm.ezfarmback.remote.domain.RemoteRepository;
-import com.ezfarm.ezfarmback.remote.dto.IotInfo;
+import com.ezfarm.ezfarmback.common.iot.IotInfo;
 import com.ezfarm.ezfarmback.remote.dto.RemoteRequest;
 import com.ezfarm.ezfarmback.remote.dto.RemoteResponse;
 import com.ezfarm.ezfarmback.user.domain.User;
@@ -61,9 +61,9 @@ public class RemoteService {
         if (findRemote.getFarm().isNotPossibleToAccessFarm(user.getId())) {
             throw new CustomException(ErrorCode.FARM_ACCESS_DENIED);
         }
-        JschObject jschObject = new JschObject(iotInfo.getHostname(), iotInfo.getUsername(), iotInfo.getPassword());
+        JschService jschService = new JschService(iotInfo.getHostname(), iotInfo.getUsername(), iotInfo.getPassword());
         findRemote.updateRemote(remoteRequest);
 
-        return jschObject.connect();
+        return jschService.updateRemote();
     }
 }
